moved auth middleware logic into its own folder
